- Константы вынесены в constants.Constants - Валидация вынесена в validation.Validation; убрана вложенность if-ов - Тесты валидации собраны в классе ValidationTest
